Add files via upload
The spreadsheet "Hydroxyapatiterecordoftests.xlsx" contains two worksheets.  The first one corresponds to the sintering temperature screening experimental conditions.  The second one contains combined pycnometry and physisorption data for a subset of hydroxyapatite/alginate formulations and serves as the foundation of the mini database under construction.  The "Filename" column contains the key to the raw data files and analysis.  The physisorption data in terms of the BET surface area has a "Physisorption" prefix in the filename.